requirements.txt
pillow
beautifulsoup4
matplotlib
requests